Dzaki

+++ app/Http/Controllers/AngkatanController.php
@@ -3,9 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Models\Angkatan;
+use App\Models\Biaya;
 use App\Models\Instansi;
 use App\Models\Jurusan;
 use App\Models\Kelas;
+use App\Models\Murid;
 use Illuminate\Http\Request;
 
 class AngkatanController extends Controller
@@ -16,7 +18,7 @@ class AngkatanController extends Controller
     public function index()
     {
         $instansi = Instansi::first();
-        $angkatan = Angkatan::all();
+        $angkatan = Angkatan::orderBy('created_at', 'desc')->get();
         return view('admin.angkatan.index', compact('angkatan', 'instansi'));
     }
 
@@ -35,7 +37,7 @@ class AngkatanController extends Controller
     public function store(Request $request)
     {
         $data = $request->validate([
-            'tahun' => 'required|unique:angkatans,tahun' 
+            'tahun' => 'required|unique:angkatans,tahun'
         ], [
             'tahun.unique' => 'Tahun angkatan sudah ada.'
         ]);
@@ -68,7 +70,7 @@ class AngkatanController extends Controller
     public function update(Request $request, string $id)
     {
         $data = $request->validate([
-            'tahun' => 'required|unique:angkatans,tahun' 
+            'tahun' => 'required|unique:angkatans,tahun'
         ], [
             'tahun.unique' => 'Tahun angkatan sudah ada.'
         ]);
@@ -77,21 +79,54 @@ class AngkatanController extends Controller
         $angkatan->update($data);
         return redirect()->route('admin.angkatan.index')->with('pesan', "Angkatan Berhasil Di Update!!");
     }
+    public function deleteSelect(Request $request)
+    {
+        $ids = $request->ids;
+        $angkatan = Angkatan::whereIn('id', $ids);
+        $biayat = Biaya::whereIn('id_angkatans', $ids)->first();
+        $murids = Murid::whereIn('id_angkatans', $ids)->first();
 
+
+        if ($biayat) {
+            foreach ($ids as $id) {
+                $biaya = Biaya::where('id_angkatans', $id)->get();
+                foreach ($biaya as $biayas) {
+                    return redirect()->route('admin.angkatan.index')->with('pesan', 'Gagal Menghapus data ! Hapus data Biaya angkatan ' . $biayas->angkatans->tahun . ' terlebih dahulu');
+                }
+            }
+        } else if ($murids) {
+            foreach ($ids as $id) {
+                $murids = Murid::where('id_angkatans', $ids)->get();
+                foreach ($murids as $murid) {
+                    return redirect()->route('admin.angkatan.index')->with('pesan', 'Gagal Menghapus data ! Hapus data murid angkatan ' . $murid->angkatans->tahun . ' terlebih dahulu');
+                }
+            }
+        } else {
+            Kelas::whereIn('id_angkatans', $ids)->delete();
+            Jurusan::whereIn('id_angkatans', $ids)->delete();
+            $angkatan->delete();
+            return redirect()->route('admin.angkatan.index')->with('success', 'success Delete data ');
+        }
+    }
     /**
      * Remove the specified resource from storage.
      */
     public function destroy(string $id)
     {
         $angkatan = Angkatan::findOrFail($id);
+        $biaya = Biaya::where('id_angkatans', $angkatan->id)->first();
+        $murids = Murid::where('id_angkatans', $angkatan->id)->first();
+        if ($biaya) {
+            return redirect()->route('admin.angkatan.index')->with('pesan', 'Gagal Menghapus data ! Hapus data Biaya angkatan ' . $biaya->angkatans->tahun . ' terlebih dahulu');
+        } else if ($murids) {
+            return redirect()->route('admin.angkatan.index')->with('pesan', 'Gagal Menghapus data ! Hapus data Murid angkatan ' . $murids->angkatans->tahun . ' terlebih dahulu');
+        } else {
+            Kelas::where('id_angkatans', $angkatan->id)->delete();
+            Jurusan::where('id_angkatans', $angkatan->id)->delete();
 
-        Kelas::where('id_angkatans', $angkatan->id)->delete();
-        Kelas::where('id_jurusans', $angkatan->id)->delete();
+            $angkatan->delete();
 
-        Jurusan::where('id_angkatans', $angkatan->id)->delete();
-
-        $angkatan->delete();
-
-        return redirect()->route('admin.angkatan.index');
+            return redirect()->route('admin.angkatan.index')->with('success', 'success Delete data ');;
+        }
     }
 }
